move api_helper to webbreaker/common and remove some fortify api declarations
