+++ webbreaker/fortify/common/fortify_helper.py
@@ -6,7 +6,7 @@ import socket
 from webbreaker.common.webbreakerhelper import WebBreakerHelper
 from webbreaker.common.webbreakerlogger import Logger
 from fortifyapi.fortify import FortifyApi
-
+from webbreaker.common.api_response_helper import APIHelper
 
 class FortifyClient(object):
     def __init__(self, fortify_url, project_template=None, application_name=None, fortify_username=None,
@@ -22,6 +22,7 @@ class FortifyClient(object):
         self.token = token
         if not token:
             self.token = self.get_token()
+            self.api = FortifyApi(self.ssc_server, token=self.token, verify_ssl=False)
 
         if not self.token:
             raise ValueError("Unable to obtain a Fortify API token.")
@@ -30,11 +31,12 @@ class FortifyClient(object):
         try:
             api = FortifyApi(self.ssc_server, username=self.user, password=self.password, verify_ssl=False)
             response = api.get_token()
-            if response.success:
-                token = response.data['data']['token']
-                return token
-            else:
-                Logger.app.critical(response.message)
+
+            APIHelper().check_for_response_fail_(response)
+
+            token = response.data['data']['token']
+            return token
+
         except Exception as e:
             if hasattr(e, 'message'):
                 Logger.app.critical("Exception while getting Fortify token: {0}".format(e.message))
@@ -42,8 +44,10 @@ class FortifyClient(object):
         return None
 
     def __get_project_id__(self, project_name):
-        api = FortifyApi(self.ssc_server, token=self.token, verify_ssl=False)
-        response = api.get_projects()
+
+        response = self.api.get_projects()
+
+        # Not replacing with APIHelper() due to some special logic going on with the return.
         if response.success:
             for project in response.data['data']:
                 if project['name'] == project_name:
@@ -61,13 +65,12 @@ class FortifyClient(object):
         Create, add required attributes to, and commit a new project version
         :return: The new project_version_id if successful. Otherwise, None.
         """
-        api = FortifyApi(self.ssc_server, token=self.token, verify_ssl=False)
         try:
-            response = api.create_project_version(project_name=self.application_name,
-                                                  project_id=self.__get_project_id__(self.application_name),
-                                                  project_template=self.project_template,
-                                                  version_name=self.fortify_version,
-                                                  description=self.__project_version_description__())
+            response = self.api.create_project_version(project_name=self.application_name,
+                                                       project_id=self.__get_project_id__(self.application_name),
+                                                       project_template=self.project_template,
+                                                       version_name=self.fortify_version,
+                                                       description=self.__project_version_description__())
 
             if not response.success:
                 raise ValueError("Failed to create a new project version")
@@ -75,7 +78,7 @@ class FortifyClient(object):
             project_version_id = response.data['data']['id']
 
             # At Target, only one attribute is required
-            response = api.add_project_version_attribute(project_version_id=project_version_id,
+            response = self.api.add_project_version_attribute(project_version_id=project_version_id,
                                                          attribute_definition_id=self.__get_attribute_definition_id__(
                                                              search_expression='name:"CI Number"'),
                                                          value='New WebBreaker Application',
@@ -83,10 +86,10 @@ class FortifyClient(object):
             if not response.success:
                 raise ValueError("Failed to create required project version attribute")
 
-            response = api.commit_project_version(project_version_id=project_version_id)
+            response = self.api.commit_project_version(project_version_id=project_version_id)
+
             if not response.success:
                 raise ValueError("Failed to commit new project version")
-                # Logger.app.debug("Created new project version id {0}".format(project_version_id))
             return project_version_id
 
         except Exception as e:
@@ -95,14 +98,12 @@ class FortifyClient(object):
         return None
 
     def __create_new_project_version__(self):
-        api = FortifyApi(self.ssc_server, token=self.token, verify_ssl=False)
-
         try:
             # It's kinda dumb for this api call to require both project name and id?
-            response = api.create_new_project_version(project_name=self.application_name,
-                                                      project_template=self.project_template,
-                                                      version_name=self.fortify_version,
-                                                      description=self.__project_version_description__())
+            response = self.api.create_new_project_version(project_name=self.application_name,
+                                                           project_template=self.project_template,
+                                                           version_name=self.fortify_version,
+                                                           description=self.__project_version_description__())
 
             if not response.success:
                 raise ValueError("Failed to create a new project version")
@@ -119,7 +120,7 @@ class FortifyClient(object):
             https://github.com/target/fortifyapi/blob/134477ef708e63d8c7b555741f5c69610abad920/fortifyapi/fortify.py#L105
             search_expression='name:"CI Number"'
             """
-            response = api.add_project_version_attribute(project_version_id=project_version_id,
+            response = self.api.add_project_version_attribute(project_version_id=project_version_id,
                                                          attribute_definition_id=self.__get_attribute_definition_id__(
                                                              search_expression='name:"CI Number"'),
                                                          value='New WebBreaker Application',
@@ -127,7 +128,7 @@ class FortifyClient(object):
             if not response.success:
                 raise ValueError("Failed to create required project version attribute")
 
-            response = api.commit_project_version(project_version_id=project_version_id)
+            response = self.api.commit_project_version(project_version_id=project_version_id)
             if not response.success:
                 raise ValueError("Failed to commit new project version")
                 # Logger.app.debug("Created new project version id {0}".format(project_version_id))
@@ -137,8 +138,7 @@ class FortifyClient(object):
         return project_version_id
 
     def __get_attribute_definition_id__(self, search_expression):
-        api = FortifyApi(self.ssc_server, token=self.token, verify_ssl=False)
-        response = api.get_attribute_definition(search_expression=search_expression)
+        response = self.api.get_attribute_definition(search_expression=search_expression)
         if response.success:
             return response.data['data'][0]['id']
         else:
@@ -151,21 +151,24 @@ class FortifyClient(object):
         If none of the above succeeds, log the reason(s) and return None
         :return:
         """
-        api = FortifyApi(self.ssc_server, token=self.token, verify_ssl=False)
         try:
-            response = api.get_all_project_versions()  # api should support a search expression here. alas...
+            response = self.api.get_all_project_versions()  # api should support a search expression here. alas...
+
             if response.success:
+
                 for project_version in response.data['data']:
                     if project_version['project']['name'] == self.application_name:
                         if project_version['name'] == self.fortify_version:
                             # we have a matching project version
                             Logger.app.debug("Found existing project version {0}".format(project_version['id']))
                             return project_version['id']
+                        
                 # Didn't find a matching project version, verify that our project exists
                 for project_version in response.data['data']:
                     if project_version['project']['name'] == self.application_name:
-                        # Our project exsits, so create a new version
+                        # Our project exists, so create a new version
                         return self.__create_project_version__()
+
                 # Let upload_scan know that our project doesn't exist
                 return -2
             elif "401" in response.message:
